Multiple changes
Removed local xcuserdata from soure control, since this is for local UI state and such. Not used for others and could change if somebody else's UI interface changes (ie, open files and xcode specific layouts).
Moved schemas for building to a shared project file. This allows for the building and targeting of the platforms. It removes the schemas from the local user settings so they are persistent across multiple builds. Also useful for a continuous build server.
Removed properties from Delegate since settings have been updated to respond to the changed event and handled via the Controller (like they are supposed to be).
Added TODO comments about the future tasks that are needed to be completed.