Fixed URL encoding error
URLs were not being properly encoded prior to usage. If a space was added to the URL, then the connection would fail because of an improperly formatted URL and the error description said "bad URL" with a "NULL" value displayed. The solution was to strip the whitespace around the beginning and end of the hostname, port and username as well as encode the URLs before use. This way if there is an improperly formatted string, the encoding will catch it. Regardless, extra spaces are silently ignored and not updated in the preferences. Could possibly add logic to the code to update the preference and remove any leading or trailing spaces. This has not been done yet.